backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.6.1
python-dotenv>=1.0.0

# LLM and Vector Store
cerebras-cloud-sdk>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Agent framework (replaced ag2)
langchain>=0.0.300

# Utilities
python-multipart>=0.0.6
httpx>=0.28.1
numpy>=1.26.0
pandas>=2.1.1

# Development
pytest==7.4.2
black==23.9.1
flake8==6.1.0
isort==5.12.0